requirements.txt
streamlit
pandas
numpy
torch
torchvision
plotly
Pillow
requests
scikit-learn
matplotlib
